+++ module/layout/index.mjs
@@ -1,3 +1,3 @@
-export * from './port.js';
-export * from './portLabel.js';
+export * from './port.mjs';
+export * from './portLabel.mjs';
 export * from '../../plugins/layout/DirectedGraph/joint.layout.DirectedGraph.js';
