+++ rollup.config.js
@@ -2,7 +2,7 @@ import * as modules from './rollup.resources';
 
 export default [
     modules.joint,
-    // modules.jointCore,
-    // modules.geometry,
-    // modules.vectorizer
+    modules.jointCore,
+    modules.geometry,
+    modules.vectorizer
 ]
